+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 
 ### Features
 
-* serverless valkey cache with a per-link user and password
+* serverless valkey cache with a per-link user, password and TLS connection url